Revert "MoveLineService structure refactor"
This reverts commit aea44f4f6b960655d46ec07626b7ff31d01ee38e.
